refactor: move lib-chainstorage to internal package
