+++ packages/inter-protocol/src/proposals/econ-behaviors.js
@@ -5,8 +5,8 @@ import '@agoric/governance/exported.js';
 import { deeplyFulfilledObject, makeTracer } from '@agoric/internal';
 import '@agoric/vats/exported.js';
 import '@agoric/vats/src/core/types.js';
-import { makeStorageNodeChild } from '@agoric/vats/src/lib-chainStorage.js';
 import { Stable, Stake } from '@agoric/vats/src/tokens.js';
+import { makeStorageNodeChild } from '@agoric/internal/src/lib-chainStorage.js';
 import { makeRatio } from '@agoric/zoe/src/contractSupport/index.js';
 import { E } from '@endo/far';
 import * as Collect from '../collect.js';
