+++ packages/internal/test/test-lib-chainStorage.js
@@ -1,7 +1,8 @@
 // @ts-check
-import { test } from '@agoric/swingset-vat/tools/prepare-test-env-ava.js';
+import test from 'ava';
+import '@endo/init';
 
-import { makeFakeStorageKit } from '../tools/storage-test-utils.js';
+import { makeFakeStorageKit } from '../src/storage-test-utils.js';
 
 test('makeChainStorageRoot', async t => {
   const rootPath = 'root';
